More Changes, trying to decide on rendering layout
