--- src/keyconfig.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "raylib.h"
-
-namespace KeyConfig 
-{
-
-static constexpr int view_mode = KEY_LEFT_SHIFT;
-static constexpr int view_mode_pan = KEY_SPACE;
-
-};
